trying to make windoze happy.
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@2431 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/libs/orcaobj/mathdefs.h
@@ -86,7 +86,7 @@ inline void NORMALISE_ANGLE( float &theta )
 {
     double thDouble = theta;
     NORMALISE_ANGLE( thDouble );
-    theta = thDouble;
+    theta = (float)thDouble;
 }
 
 /*****************************************************************************
